Update code:
I removed the distributed lock since we are going to assume that we have pre-initialized the counter.
Updated the driver (Main) to have spring actuator - I had to change up some of the code to get actuator to work.
Threw in a little "test" class that resets the counter to zero.